// Assets-Uncapped/front-portal-develop/src/services/api/organisation-users/models/ExternalCompanyResponse.ts
export interface ExternalCompanyResponse {
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    companyNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    companyStatus?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    addressSnippet?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    usState?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    region?: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalCompanyResponse
     */
    website?: string;
    /**
     * 
     * @type {Date}
     * @memberof ExternalCompanyResponse
     */
    dateOfCreation?: Date;
    /**
     * 
     * @type {CompanyAddressDTO}
     * @memberof ExternalCompanyResponse
     */
    fullAddress?: CompanyAddressDTO;
}